Refactor tutorials to support multiple languages.
The way it works is we define an entry point for each tutorial with some metadata (description and product) under config/tutorials/en/, e.g.

config/tutorials/en/client-sdk/phone-to-app.md
and the different code languages should live in a folder under the same name:

config/
  tutorials/
    en/
      client-sdk/
        phone-to-app/
          javascript.yml
          python.yml

the different steps are defined as before in _tutorials/en/client-sdk with the caveat that now we can have shared steps across different code languages or have specific definitions for them, .i.e.

_tutorials/en/client-sdk/
  phone-to-app/
    code/ <- specific to each code language
      python.md
      javascript.md
    create-application.md <-shared
    create-conversation.md <- shared

What is included:

* Rename files in order to match the new configuration.
* Display a list of the available languages for a tutorial
* Update `/tutorials` to include the different languages a tutorial is
available and show its product. In case of multiple languages, the
description and title of the entry point is displayed.
* Refactor how TutorialList work
* Simplified TutorialsController#index
* Add Product.rb

What needs to be done:
* Fix and add more specs.
* Refactor Tutorial, TutorialList and TutorialListItem.